refactor(core): split commands + secrets vault (<1000 lines)
- commands.rs (1289) → commands/ dir: the /skill, /thread, /clawhub
  subcommand handlers move to commands/subcommands.rs; mod.rs keeps the
  CommandAction/Context types and the main handle_command dispatch.
- secrets/vault.rs (1157): split the SecretsManager impl across vault.rs +
  vault_ext.rs (credential management methods) as a second impl block.

No behavior change.
